highperfhooks: add option to configure irqbalance restore
Since #4441 , to support
the dynamic irqbalance configuration in turn needed by the high
performance hooks, CRI-O runs the restore code at startup.

In some configuration scenarios it's preferred to have
more control over this step. Some configuration wants
to have more control over the mask to be restored, or to
disable the restore flow entirely.

For this reason, we add an option to provide a user-supplied
restore file. Users wishing to exert more control can provide
their data, or they can disable the flow entirely setting
the path to the restore file to empty string ("").

The new option defaults are meant to ensure full backward
compatibility.

Signed-off-by: Francesco Romani <fromani@redhat.com>